ml-service/requirements.txt
# FastAPI ve Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# HTTP Client
httpx==0.25.2
requests==2.31.0

# ML ve Data Science
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.25.2
matplotlib==3.8.2
plotly==5.17.0

# H3 ve Geospatial Libraries
h3==4.1.1

# Redis ve Message Queue
redis==5.0.1
pika==1.3.2

# HTTP Client (for OpenStreetMap)
httpx==0.25.2

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
aiofiles==23.2.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development
black==23.11.0
flake8==6.1.0 